Fix for https://issues.apache.org/bugzilla/show_bug.cgi?id=42707 I debated adding the addContainerListener() meethod to JMX. I'll leave it a little while for people to comment on this (or anything else) before I propose this one for 6.0.x.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@712467 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/connector/MapperListener.java
@@ -26,6 +26,10 @@ import javax.management.NotificationListener;
 import javax.management.ObjectInstance;
 import javax.management.ObjectName;
 
+import org.apache.catalina.ContainerEvent;
+import org.apache.catalina.ContainerListener;
+import org.apache.catalina.Host;
+import org.apache.catalina.ServerFactory;
 import org.apache.catalina.core.StandardContext;
 import org.apache.juli.logging.Log;
 import org.apache.juli.logging.LogFactory;
@@ -46,8 +50,10 @@ import org.apache.tomcat.util.res.StringManager;
  * @author Costin Manolache
  */
 public class MapperListener
-    implements NotificationListener 
+    implements NotificationListener, ContainerListener
  {
+
+
     private static Log log = LogFactory.getLog(MapperListener.class);
 
 
@@ -254,6 +260,19 @@ public class MapperListener
     }
 
 
+    // --------------------------------------------- Container Listener methods
+
+    public void containerEvent(ContainerEvent event) {
+
+        if (event.getType() == Host.ADD_ALIAS_EVENT) {
+            mapper.addHostAlias(((Host) event.getSource()).getName(),
+                    event.getData().toString());
+        } else if (event.getType() == Host.REMOVE_ALIAS_EVENT) {
+            mapper.removeHostAlias(event.getData().toString());
+        }
+    }
+
+    
     // ------------------------------------------------------ Protected Methods
 
     private void registerEngine()
@@ -296,7 +315,7 @@ public class MapperListener
             if (!isRegisteredWithAlias && log.isWarnEnabled())
                 log.warn(sm.getString("mapperListener.unknownDefaultHost", defaultHost));
         }
-        // This should probablt be called later 
+        // This should probably be called later 
         if( defaultHost != null ) {
             mapper.setDefaultHostName(defaultHost);
         }
@@ -309,13 +328,16 @@ public class MapperListener
         throws Exception {
         String name=objectName.getKeyProperty("host");
         if( name != null ) {        
-            String[] aliases = (String[])
-                mBeanServer.invoke(objectName, "findAliases", null, null);
+
+            Host host = (Host) ServerFactory.getServer().findService(
+                    domain).getContainer().findChild(name);
+        
+            String[] aliases = host.findAliases();
             mapper.addHost(name, aliases, objectName);
+            host.addContainerListener(this);
             if(log.isDebugEnabled())
                 log.debug(sm.getString
                      ("mapperListener.registerHost", name, domain));
-
         }
     }
 
@@ -326,10 +348,16 @@ public class MapperListener
     private void unregisterHost(ObjectName objectName)
         throws Exception {
         String name=objectName.getKeyProperty("host");
-        mapper.removeHost(name);
-        if(log.isDebugEnabled())
-            log.debug(sm.getString
-                 ("mapperListener.unregisterHost", name, domain));
+        if( name != null ) { 
+            Host host = (Host) ServerFactory.getServer().findService(
+                domain).getContainer().findChild(name);
+        
+            mapper.removeHost(name);
+            host.removeContainerListener(this);
+            if(log.isDebugEnabled())
+                log.debug(sm.getString
+                        ("mapperListener.unregisterHost", name, domain));
+        }
     }
